+++ solutions/uri/1026/1026.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
-int main() {
-    unsigned int a, b;
+int16_t main() {
+    u_int16_t a, b;
 
     while (scanf("%u %u", &a, &b) != EOF) {
         printf("%u\n", a ^ b);
